requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# File Processing
unstructured==0.11.2
pdfplumber==0.10.3
PyMuPDF==1.23.8
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
beautifulsoup4==4.12.2
lxml==4.9.3

# OCR
pytesseract==0.3.10
Pillow==10.1.0

# AI/LLM
openai>=1.6.1
anthropic>=0.8.0
langchain>=0.1.0
langchain-openai>=0.0.5

# Database & Storage
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
boto3==1.34.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
aiofiles==23.2.1

# Security & Validation
python-magic==0.4.27
validators==0.22.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging & Monitoring
structlog==23.2.0